requirements.txt
pandas>=1.5.0
numpy>=1.23.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-dotenv>=1.0.0

